+++ osidb/tests/test_embargo.py
@@ -10,31 +10,48 @@ pytestmark = pytest.mark.unit
 
 
 class TestEmbargo(object):
-    def test_factory(self, public_groups, embargoed_groups):
+    def test_factory(
+        self,
+        public_read_groups,
+        public_write_groups,
+        embargoed_read_groups,
+        embargoed_write_groups,
+    ):
         flaw = FlawFactory(embargoed=False)
-        assert flaw.acl_read == public_groups
+        assert flaw.acl_read == public_read_groups
+        assert flaw.acl_write == public_write_groups
 
         flaw = FlawFactory(embargoed=True)
-        assert flaw.acl_read == embargoed_groups
+        assert flaw.acl_read == embargoed_read_groups
+        assert flaw.acl_write == embargoed_write_groups
 
     @pytest.mark.parametrize("embargoed", [False, True])
     @freeze_time(timezone.datetime(2022, 11, 25))
-    def test_embargoed_annotation(self, embargoed_groups, public_groups, embargoed):
+    def test_embargoed_annotation(
+        self,
+        public_read_groups,
+        public_write_groups,
+        embargoed_read_groups,
+        embargoed_write_groups,
+        embargoed,
+    ):
         if embargoed:
-            groups = embargoed_groups
+            read_groups = embargoed_read_groups
+            write_groups = embargoed_write_groups
             title = "EMBARGOED CVE-2022-1234 kernel: some description"
             unembargo_dt = timezone.datetime(
                 2022, 12, 26, tzinfo=timezone.get_current_timezone()
             )
         else:
-            groups = public_groups
+            read_groups = public_read_groups
+            write_groups = public_write_groups
             title = "CVE-2022-1234 kernel: some description"
             unembargo_dt = timezone.datetime(
                 2022, 11, 24, tzinfo=timezone.get_current_timezone()
             )
         flaw = Flaw(
-            acl_read=groups,
-            acl_write=groups,
+            acl_read=read_groups,
+            acl_write=write_groups,
             cve_id="CVE-2000-11111",
             cwe_id="CWE-1",
             type="VULNERABILITY",
@@ -52,11 +69,13 @@ class TestEmbargo(object):
         assert flaw.embargoed == embargoed
 
     @pytest.mark.parametrize("embargoed", [False, True])
-    def test_embargoed_readonly(self, public_groups, embargoed_groups, embargoed):
+    def test_embargoed_readonly(
+        self, public_read_groups, public_write_groups, embargoed
+    ):
         with pytest.raises(TypeError) as ex:
             flaw = Flaw(
-                acl_read=public_groups,
-                acl_write=public_groups,
+                acl_read=public_read_groups,
+                acl_write=public_write_groups,
                 cve_id="CVE-2000-11111",
                 type="VULNERABILITY",
                 state="NEW",
@@ -70,66 +89,3 @@ class TestEmbargo(object):
             )
             flaw.save()
         assert "Flaw() got an unexpected keyword argument 'embargoed'" in str(ex)
-
-    # TODO the following is not applicable any more
-    # but should be resurrected once we start implement write functionality
-    #
-    # def test_setting_embargo(self):
-    #     """explicitly test setting of embargo"""
-    #     flaw = FlawFactory()
-
-    #     # | embargoed | unembargo_dt | embargoed set value |
-    #     # |-----------|--------------|---------------------|
-    #     # | True      | None         | True                |
-    #     flaw.embargoed = True
-    #     flaw.unembargo_dt = None
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is True
-
-    #     # | False      | None        | False               |
-    #     flaw.embargoed = False
-    #     flaw.unembargo_dt = None
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is False
-
-    #     # | None      | None         | False               |
-    #     flaw.embargoed = None
-    #     flaw.unembargo_dt = None
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is False
-
-    #     # | True      | Future date  | True                |
-    #     flaw.embargoed = True
-    #     flaw.unembargo_dt = datetime.now() + timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is True
-
-    #     # | False     | Future date  | True                 |
-    #     flaw.embargoed = False
-    #     flaw.unembargo_dt = datetime.now() + timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is True
-
-    #     # | None      | Future date  | True                 |
-    #     flaw.embargoed = None
-    #     flaw.unembargo_dt = datetime.now() + timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is True
-
-    #     # | True      | Past date    | True                |
-    #     flaw.embargoed = True
-    #     flaw.unembargo_dt = datetime.now() - timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is True
-
-    #     # | False     | Past date    | False                |
-    #     flaw.embargoed = False
-    #     flaw.unembargo_dt = datetime.now() - timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is False
-
-    #     # | None      | Past date    | False                |
-    #     flaw.embargoed = None
-    #     flaw.unembargo_dt = datetime.now() - timedelta(days=1)
-    #     assert flaw.process_embargo_state() is None
-    #     assert flaw.embargoed is False